+++ src/main/scala/com/github/unchama/seichiassist/subsystems/breakcount/subsystems/notification/bukkit/actions/BukkitNotifyLevelUp.scala
@@ -12,22 +12,22 @@ import com.github.unchama.seichiassist.subsystems.breakcount.domain.level.{
 }
 import com.github.unchama.seichiassist.subsystems.breakcount.subsystems.notification.application.actions.NotifyLevelUp
 import com.github.unchama.seichiassist.util.{
+  LunchFireWorksEffect,
   PlayerSendable,
   SendMessageEffect,
-  SendSoundEffect,
-  Util
+  SendSoundEffect
 }
 import org.bukkit.ChatColor.{BOLD, GOLD}
-import org.bukkit.entity.Player
 import org.bukkit.Sound
+import org.bukkit.entity.Player
 
 //FIXME ファイル名とやっていることが違うようになっているので修正するべき。
 //例えば、10億の倍数到達時の通知はLevelUp時の通知ではない
 //また、BukkitNotifyLevelUpなのにdiffの展開やいつメッセージを出すかなどを扱うべきでない。
 object BukkitNotifyLevelUp {
 
-  import cats.implicits._
   import PlayerSendable.forString
+  import cats.implicits._
 
   def apply[F[_]: OnMinecraftServerThread: Sync]: NotifyLevelUp[F, Player] =
     new NotifyLevelUp[F, Player] {
@@ -60,7 +60,7 @@ object BukkitNotifyLevelUp {
           OnMinecraftServerThread[F].runAction(SyncIO {
             player.sendTitle(titleMessage, subtitleMessage, 1, 20 * 5, 1)
             player.sendMessage(s"$subtitleMessage$titleMessage")
-            Util.launchFireWorks(player.getLocation)
+            LunchFireWorksEffect.launchFireWorks(player.getLocation)
           })
         } else Applicative[F].unit
       }
@@ -79,7 +79,7 @@ object BukkitNotifyLevelUp {
         if (oldStars < newStars) Sync[F].delay {
           player.sendTitle(titleMessage, subTitleMessage, 1, 20 * 5, 1)
           player.sendMessage(s"$subTitleMessage$titleMessage")
-          Util.launchFireWorks(player.getLocation)
+          LunchFireWorksEffect.launchFireWorks(player.getLocation)
         }
         else Applicative[F].unit
       }
