+++ src/main/scala/com/github/unchama/seichiassist/util/Util.scala
@@ -4,71 +4,18 @@ import org.bukkit.ChatColor._
 import org.bukkit._
 import org.bukkit.block.{Block, Skull}
 import org.bukkit.entity.EntityType._
-import org.bukkit.entity.{EntityType, Firework, LivingEntity, Player}
+import org.bukkit.entity.{EntityType, LivingEntity, Player}
 import org.bukkit.event.entity.EntityDamageEvent.DamageCause
 import org.bukkit.inventory.meta.SkullMeta
 import org.bukkit.inventory.{ItemFlag, ItemStack}
 
+import java.util.Calendar
 import java.util.stream.IntStream
-import java.util.{Calendar, Random}
 
 object Util {
 
   import scala.jdk.CollectionConverters._
   import scala.util.chaining._
-
-  private val types = List(
-    FireworkEffect.Type.BALL,
-    FireworkEffect.Type.BALL_LARGE,
-    FireworkEffect.Type.BURST,
-    FireworkEffect.Type.CREEPER,
-    FireworkEffect.Type.STAR
-  )
-
-  // 指定された場所に花火を打ち上げる関数
-  def launchFireWorks(loc: Location): Unit = {
-    // 花火を作る
-    val firework = loc.getWorld.spawn(loc, classOf[Firework])
-
-    // 花火の設定情報オブジェクトを取り出す
-    val meta = firework.getFireworkMeta
-    val effect = FireworkEffect.builder()
-    val rand = new Random()
-
-    // 形状をランダムに決める
-    effect.`with`(types(rand.nextInt(types.size)))
-
-    // 基本の色を単色～5色以内でランダムに決める
-    effect.withColor(getRandomColors(1 + rand.nextInt(5)): _*)
-
-    // 余韻の色を単色～3色以内でランダムに決める
-    effect.withFade(getRandomColors(1 + rand.nextInt(3)): _*)
-
-    // 爆発後に点滅するかをランダムに決める
-    effect.flicker(rand.nextBoolean())
-
-    // 爆発後に尾を引くかをランダムに決める
-    effect.trail(rand.nextBoolean())
-
-    // 打ち上げ高さを1以上4以内でランダムに決める
-    meta.setPower(1 + rand.nextInt(4))
-
-    // 花火の設定情報を花火に設定
-    meta.addEffect(effect.build())
-
-    firework.setFireworkMeta(meta)
-  }
-
-  // カラーをランダムで決める
-  def getRandomColors(length: Int): Array[Color] = {
-    // 配列を作る
-    val rand = new Random()
-    // 配列の要素を順に処理していく
-    // 24ビットカラーの範囲でランダムな色を決める
-
-    // 配列を返す
-    (0 until length).map { _ => Color.fromBGR(rand.nextInt(1 << 24)) }.toArray
-  }
 
   def isGachaTicket(itemStack: ItemStack): Boolean = {
     val containsRightClickMessage: String => Boolean = _.contains(s"${GREEN}右クリックで使えます")
